lint-based updates.

+++ src.d/logout_user.c
@@ -32,7 +32,7 @@ int main(int argc, char *argv[])
       (void) memset(tmp2, 0, sizeof(tmp2));
       (void) snprintf(tmp2, sizeof(tmp2), "%s", argv[1]);
 
-      for(i = 0; i < strlen(tmp2); i++)
+      for(i = 0; i < (int) strlen(tmp2); i++)
 	if(tmp2[i] == '-')
 	  {
 	    tmp2[i] = '\0';
